Re-worked syntax for starting up new Xserver
This is an incompatible change. The previous solution was problematic
in that spelling errors, wrong order etc could be mistaken as a
request to start a new Xserver. Since :0 is default, the result was
often unexpected.

With the new syntax, the 2nd form is only used if --start is
given. This means that we again allow parameters after the server
display; just like vncviewer.

This is an update to 3ec76bd5eb9d70ba3dde7dcfd5c5e0aa3d7d3617.

Also added a test script.